# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NovaChat Canonical HTTP API
  version: 1.0.0
  description: |
    Source-controlled contract for the authenticated NovaChat web client.
    Cassandra remains authoritative; cursors are opaque and must not be decoded by clients.
servers:
  - url: http://localhost:8084/api
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /files/upload:
    post:
      tags: [Files]
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary}}}}}}
      responses: {'201': {description: Uploaded asset}}
  /files/upload/multiple:
    post:
      tags: [Files]
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [files], properties: {files: {type: array, maxItems: 10, items: {type: string, format: binary}}}}}}}
      responses: {'201': {description: Uploaded assets}}
  /files/{assetId}:
    delete:
      tags: [Files]
      parameters: [{name: assetId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'204': {description: Asset deleted}}
...
  /reports:
    post:
      tags: [Moderation]
      summary: Submit a user report
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateReport'}}}}
      responses: {'201': {description: Created report}}
  /reports/mine:
    get:
      tags: [Moderation]
      parameters: [{$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Reports submitted by the authenticated user}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  parameters:
...
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
  schemas:
...
    CreateReport: {type: object, required: [targetType, reasonCode], properties: {targetType: {type: string, enum: [USER, MESSAGE, CONVERSATION]}, targetUserId: {type: string, format: uuid}, conversationId: {type: string, format: uuid}, messageBucket: {type: string}, messageId: {type: string, format: uuid}, reasonCode: {type: string, maxLength: 80}, description: {type: string, maxLength: 2000}}}
...